added: space setting, space amenities sub modules plus query to publish space in space module

+++ src/functions/graphql/schema/space/removeSpace.ts
@@ -20,9 +20,11 @@ const removeSpace: RemoveSpace = async (_, { id }, { authData, dataSources, stor
     if (accountId !== space.accountId)
         throw new GqlError({ code: "FORBIDDEN", message: "You are not allowed to modify this space" });
 
-    await store.space.update({ where: { id }, data: { isDeleted: true } });
+    const deletedSpace = await store.space.update({ where: { id }, data: { isDeleted: true } });
 
-    await dataSources.spaceAlgolia.deleteObject(id);
+    if (deletedSpace.published) {
+        await dataSources.spaceAlgolia.deleteObject(id);
+    }
 
     return { message: `Successfully removed space named ${space.name}` };
 };
